% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceMeasure.R
\name{InfluenceMeasure}
\alias{InfluenceMeasure}
\title{InfluenceMeasure}
\usage{
InfluenceMeasure(data, measure, output = "plot")
}
\arguments{
\item{data}{either a lm-object or a dataframe, the code expect to both have variables named x and y.}

\item{measure}{is a string to indicate the type of influence measure to be used:
"Cooks","DFITS" or "Hadi"}

\item{output}{is string which describes the type of output the user prefers "plot" or "influences".}
}
\value{
The functions returns either plot or the influence measures.
}
\description{
The function InfluenceMeasure is aimed to identify points of high influence,which
might have a large impact in linear regression.
}
\examples{
slope <-  4
offset <- -3
sds <- 7
x <- seq(from = 1, to = 10, length.out= 100)
y <- slope*x + offset
noise <- rnorm(length(x),sd=2)
y <- y + noise
num <- 3
s <- sample(length(x), num)
y[s] <- y[s] +  rnorm(num, sd=sds) + runif(num, -sds,sds)


lin_model <- lm(y~x)
results <- InfluenceMeasure(lin_model, measure="Hadi", output = "values")

}
